Conflicts:
	RecordTimer.py
	SleepTimer.py
	data/keymap.xml
	data/menu.xml
	data/radio-hd.mvi
	data/radio.mvi
	data/skin_default/icons/Makefile.am
	data/startwizard.xml
	data/tutorialwizard.xml
	lib/python/Components/Renderer/ChannelNumber.py
	lib/python/Plugins/DemoPlugins/TestPlugin/plugin.py
	lib/python/Plugins/Extensions/DVDBurn/DVDProject.py
	lib/python/Plugins/Extensions/DVDBurn/DreamboxDVD.ddvdp.xml
	lib/python/Plugins/Extensions/DVDBurn/TitleList.py
	lib/python/Plugins/Extensions/DVDBurn/meta/plugin_dvdburn.xml
	lib/python/Plugins/Extensions/DVDPlayer/meta/plugin_dvdplayer.xml
	lib/python/Plugins/Extensions/MediaPlayer/plugin.py
	lib/python/Plugins/SystemPlugins/CleanupWizard/cleanupwizard.xml
	lib/python/Plugins/SystemPlugins/CrashlogAutoSubmit/plugin.py
	lib/python/Plugins/SystemPlugins/NFIFlash/downloader.py
	lib/python/Plugins/SystemPlugins/NetworkWizard/networkwizard.xml
	lib/python/Plugins/SystemPlugins/OldSoftwareUpdate/plugin.py
	lib/python/Plugins/SystemPlugins/SkinSelector/meta/plugin_skinselector.xml
	lib/python/Plugins/SystemPlugins/SoftwareManager/imagewizard.xml
	lib/python/Plugins/SystemPlugins/SoftwareManager/meta/plugin_softwaremanager.xml
	lib/python/Plugins/SystemPlugins/SoftwareManager/plugin.py
	lib/python/Plugins/SystemPlugins/SoftwareUpdate/plugin.py
	lib/python/Screens/DefaultWizard.py
	lib/python/Screens/NetworkSetup.py
	lib/python/Screens/SleepTimerEdit.py
	lib/python/Screens/TaskView.py
	lib/python/Tools/Downloader.py
	po/bg.po
	po/de.po

+++ lib/python/Components/ParentalControl.py
@@ -177,7 +177,7 @@ class ParentalControl:
 		else:
 			self.checkPinInterval = True
 			iMinutes = float(self.storeServicePin)
-			iSeconds = iMinutes*60
+			iSeconds = int(iMinutes*60)
 			self.pinIntervalSeconds = iSeconds
 	
 		self.configInitialized = True
